requirements.txt
tensorflow-gpu==2.0.0
numpy==1.19.1
xgboost==1.1.1
seaborn==0.10.1
scikit-learn==0.23.2
pandas==1.0.5
matplotlib==3.2.2